+++ Sources/PixelSubagent/PixelSubagent.swift
@@ -1,0 +1,3 @@
+public enum PixelSubagent {
+    public static let version = "0.2.5"
+}